pyproject.toml
[project]
name = "ganttchart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
"flask",
"requests",
"plotly",
"pandas",
"python-dotenv",
"matplotlib", "reportlab"
]
